+++ flutter_sheet_localization/lib/flutter_sheet_localization.dart
@@ -3,5 +3,6 @@ library flutter_sheet_localization;
 class SheetLocalization {
   final String docId;
   final String sheetId;
-  const SheetLocalization(this.docId, this.sheetId);
+  final int version;
+  const SheetLocalization(this.docId, this.sheetId, this.version);
 }