+++ flutter_sheet_localization_generator/lib/src/builder.dart
@@ -24,39 +24,68 @@ class DartBuilder {
 
     var constructor = ConstructorBuilder();
 
-    result.fields.add(Field((b) => b
-      ..name = "locale"
-      ..modifier = FieldModifier.final$
-      ..type = refer("Locale")));
+    result.fields.add(
+      Field(
+        (b) => b
+          ..name = 'locale'
+          ..modifier = FieldModifier.final$
+          ..type = refer('Locale'),
+      ),
+    );
 
-    constructor.requiredParameters.add(Parameter((p) => p
-      ..name = "locale"
-      ..toThis = true));
+    constructor.requiredParameters.add(
+      Parameter(
+        (p) => p
+          ..name = 'locale'
+          ..toThis = true,
+      ),
+    );
 
-    constructor.initializers.add(Code("this.labels = languages[locale]"));
+    constructor.initializers.add(
+      Code('this.labels = languages[locale]'),
+    );
 
-    result.fields.add(Field((b) => b
-      ..name = "languages"
-      ..static = true
-      ..assignment = _createLanguageMap(localizations)
-      ..modifier = FieldModifier.final$
-      ..type = refer("Map<Locale," + localizations.normalizedName + ">")));
+    result.fields.add(
+      Field(
+        (b) => b
+          ..name = 'languages'
+          ..static = true
+          ..assignment = _createLanguageMap(localizations)
+          ..modifier = FieldModifier.final$
+          ..type = refer(
+            'Map<Locale,' + localizations.normalizedName + '>',
+          ),
+      ),
+    );
 
-    result.fields.add(Field((b) => b
-      ..name = "labels"
-      ..modifier = FieldModifier.final$
-      ..type = refer(localizations.normalizedName)));
+    result.fields.add(
+      Field(
+        (b) => b
+          ..name = 'labels'
+          ..modifier = FieldModifier.final$
+          ..type = refer(localizations.normalizedName),
+      ),
+    );
 
-    result.methods.add(Method((b) => b
-      ..name = "of"
-      ..lambda = true
-      ..static = true
-      ..returns = refer(localizations.normalizedName)
-      ..body = Code(
-          "Localizations.of<${localizations.name}>(context, ${localizations.name})?.labels")
-      ..requiredParameters.add(Parameter((b) => b
-        ..name = "context"
-        ..type = refer("BuildContext")))));
+    result.methods.add(
+      Method(
+        (b) => b
+          ..name = 'of'
+          ..lambda = true
+          ..static = true
+          ..returns = refer(localizations.normalizedName)
+          ..body = Code(
+            'Localizations.of<${localizations.name}>(context, ${localizations.name})?.labels',
+          )
+          ..requiredParameters.add(
+            Parameter(
+              (b) => b
+                ..name = 'context'
+                ..type = refer('BuildContext'),
+            ),
+          ),
+      ),
+    );
 
     result.constructors.add(constructor.build());
 
@@ -66,25 +95,25 @@ class DartBuilder {
   Code _createLanguageMap(Localizations localizations) {
     final result = StringBuffer();
 
-    result.write("{");
+    result.write('{');
 
     for (var languageCode in localizations.supportedLanguageCodes) {
       final code = _createSectionInstance(languageCode, localizations);
 
-      final splits = languageCode.split("-");
+      final splits = languageCode.split('-');
 
-      var key = "Locale.fromSubtags(languageCode: \"${splits.first}\"";
+      var key = 'Locale.fromSubtags(languageCode: \"${splits.first}\"';
       if (splits.length > 2) {
-        key += ", scriptCode: \"${splits[1]}\"";
-        key += ", countryCode: \"${splits[2]}\"";
+        key += ', scriptCode: \"${splits[1]}\"';
+        key += ', countryCode: \"${splits[2]}\"';
       } else if (splits.length > 1) {
-        key += ", countryCode: \"${splits[1]}\"";
+        key += ', countryCode: \"${splits[1]}\"';
       }
-      key += ")";
-      result.write(key + " : " + code + ",");
+      key += ')';
+      result.write(key + ' : ' + code + ',');
     }
 
-    result.write("}");
+    result.write('}');
 
     return Code(result.toString());
   }
@@ -97,17 +126,23 @@ class DartBuilder {
       if (templatedValues.isNotEmpty) {
         for (var templatedValue in templatedValues) {
           if (templatedValue.type == 'DateTime') {
-            if (condition != null) {
-              value = value.replaceFirst(templatedValue.value,
-                  "\$\{DateFormat('$condition', '$languageCode').format(${templatedValue.normalizedKey})\}");
+            if (templatedValue.formatting != null) {
+              value = value.replaceFirst(
+                templatedValue.value,
+                "\$\{DateFormat('${templatedValue.formatting}', '$languageCode').format(${templatedValue.normalizedKey})\}",
+              );
             } else {
-              value = value.replaceFirst(templatedValue.value,
-                  "\$\{${templatedValue.normalizedKey}.toIso8601String()\}");
+              value = value.replaceFirst(
+                templatedValue.value,
+                '\$\{${templatedValue.normalizedKey}.toIso8601String()\}',
+              );
             }
-          } else if (templatedValue.type == 'double' ||
-              templatedValue.type == 'int' ||
-              templatedValue.type == 'num') {
-            if (condition != null) {
+          } else if (const [
+            'double',
+            'int',
+            'num',
+          ].contains(templatedValue.type)) {
+            if (templatedValue.formatting != null) {
               if ([
                 'decimalPercentPattern',
                 'currency',
@@ -116,67 +151,69 @@ class DartBuilder {
                 'compactLong',
                 'compactSimpleCurrency',
                 'compactCurrency'
-              ].contains(condition)) {
+              ].contains(templatedValue.formatting)) {
                 value = value.replaceFirst(templatedValue.value,
-                    "\$\{NumberFormat.$condition(locale: '$languageCode').format(${templatedValue.normalizedKey})\}");
+                    "\$\{NumberFormat.${templatedValue.formatting}(locale: '$languageCode').format(${templatedValue.normalizedKey})\}");
               } else if ([
                 'decimalPattern',
                 'percentPattern',
                 'scientificPattern'
-              ].contains(condition)) {
+              ].contains(templatedValue.formatting)) {
                 value = value.replaceFirst(templatedValue.value,
-                    "\$\{NumberFormat.$condition('$languageCode').format(${templatedValue.normalizedKey})\}");
-              } else if (condition == 'format') {
+                    "\$\{NumberFormat.${templatedValue.formatting}('$languageCode').format(${templatedValue.normalizedKey})\}");
+              } else if (templatedValue.formatting == 'format') {
                 value = value.replaceFirst(templatedValue.value,
-                    "\$\{NumberFormat.$condition('$languageCode').format(${templatedValue.normalizedKey})\}");
+                    "\$\{NumberFormat.${templatedValue.formatting}('$languageCode').format(${templatedValue.normalizedKey})\}");
               }
             } else {
-              value = value.replaceFirst(templatedValue.value,
-                  "\$\{NumberFormat(null, '$languageCode').format(${templatedValue.normalizedKey})\}");
+              value = value.replaceFirst(
+                templatedValue.value,
+                "\$\{NumberFormat(null, '$languageCode').format(${templatedValue.normalizedKey})\}",
+              );
             }
           } else {
             value = value.replaceFirst(
-                templatedValue.value, "\$\{${templatedValue.normalizedKey}\}");
+                templatedValue.value, '\$\{${templatedValue.normalizedKey}\}');
           }
         }
       }
-      return "\"" + _excapeString(value) + "\"";
+      return '\"' + _excapeString(value) + '\"';
     };
 
     result.write(section.normalizedName);
-    result.write("(");
+    result.write('(');
 
     section.labels.forEach((x) {
       result.write(x.normalizedKey);
-      result.write(":");
+      result.write(':');
 
       if (x.cases.length == 1 && x.cases.first.condition is DefaultCondition) {
         // Single case
         final translation = x.cases.first.translations.firstWhere(
             (x) => x.languageCode == languageCode,
-            orElse: () => Translation(languageCode, "<?" + x.key + "?>"));
+            orElse: () => Translation(languageCode, '<?' + x.key + '?>'));
 
         if (x.templatedValues.isEmpty) {
-          result.write("\"" + _excapeString(translation.value) + "\"");
+          result.write('\"' + _excapeString(translation.value) + '\"');
         } else {
           final functionArgs =
-              x.templatedValues.map((x) => x.normalizedKey).join(", ");
+              x.templatedValues.map((x) => x.normalizedKey).join(', ');
 
           // We replace all occurences of `{{original_key}}` by `$originalKey`
-          result.write("(\{$functionArgs\}) => " +
+          result.write('(\{$functionArgs\}) => ' +
               templatedString(translation.value, x.templatedValues));
         }
       } else {
         // Multiple cases
-        var functionArgs = "condition";
+        var functionArgs = 'condition';
 
         if (x.templatedValues.isNotEmpty) {
-          functionArgs += ", {" +
-              x.templatedValues.map((x) => x.normalizedKey).join(", ") +
-              "}";
+          functionArgs += ', {' +
+              x.templatedValues.map((x) => x.normalizedKey).join(', ') +
+              '}';
         }
 
-        result.write("($functionArgs) {");
+        result.write('($functionArgs) {');
 
         final categoryConditions =
             x.cases.where((x) => x.condition is CategoryCondition).toList();
@@ -184,16 +221,18 @@ class DartBuilder {
           for (var oneCase in categoryConditions) {
             final condition = oneCase.condition as CategoryCondition;
             result.write(
-                "if(condition == ${condition.category.normalizedKey}.${condition.value})");
+                'if(condition == ${condition.category.normalizedKey}.${condition.value})');
 
             final translation = oneCase.translations.firstWhere(
                 (x) => x.languageCode == languageCode,
-                orElse: () => Translation(languageCode, "<?" + x.key + "?>"));
+                orElse: () => Translation(languageCode, '<?' + x.key + '?>'));
 
-            result.write("return " +
-                templatedString(
-                    translation.value, x.templatedValues, condition.value) +
-                ";");
+            result.write(
+              'return ' +
+                  templatedString(
+                      translation.value, x.templatedValues, condition.value) +
+                  ';',
+            );
           }
         }
 
@@ -204,55 +243,58 @@ class DartBuilder {
         if (defaultCase != null) {
           final translation = defaultCase.translations.firstWhere(
               (x) => x.languageCode == languageCode,
-              orElse: () => Translation(languageCode, "<?" + x.key + "?>"));
+              orElse: () => Translation(languageCode, '<?' + x.key + '?>'));
 
-          result.write("return " +
+          result.write('return ' +
               templatedString(translation.value, x.templatedValues) +
-              ";");
+              ';');
         } else {
-          result.write("throw Exception();");
+          result.write('throw Exception();');
         }
 
-        result.write("}");
+        result.write('}');
       }
 
-      result.write(",");
+      result.write(',');
     });
 
     section.children.forEach((child) {
       result.write(child.normalizedKey);
-      result.write(":");
+      result.write(':');
       result.write(_createSectionInstance(languageCode, child));
-      result.write(",");
+      result.write(',');
     });
 
-    result.write(")");
+    result.write(')');
 
     return result.toString();
   }
 
-  String _excapeString(String value) => value.replaceAll("\"", "\\\"");
+  String _excapeString(String value) => value.replaceAll('\"', '\\\"');
 
   void _addCategoryDefinition(Category category) {
-    final values = category.values.map((x) => x + ",").join();
-    final code = Code("enum ${category.normalizedKey} { $values }");
+    final values = category.values.map((x) => x + ',').join();
+    final code = Code('enum ${category.normalizedKey} { $values }');
     _library.body.add(code);
   }
 
   String _getLabelsCase(List<Label> labels) {
-    final results = StringBuffer("switch (key) {");
+    final results = StringBuffer('switch (key) {');
     labels.forEach((label) {
-      if (label.templatedValues.isEmpty && label.cases.length == 1 && label.cases.first.condition is DefaultCondition) {
+      if (label.templatedValues.isEmpty &&
+          label.cases.length == 1 &&
+          label.cases.first.condition is DefaultCondition) {
         results.write("case '");
         results.write(label.normalizedKey);
         results.write("'");
-        results.write(":");
-        results.write("return ");
+        results.write(':');
+        results.write('return ');
         results.write(label.normalizedKey);
-        results.write(";\n");
-    }});
+        results.write(';\n');
+      }
+    });
     results.write("default: return '';");
-    results.write("}");
+    results.write('}');
     return results.toString();
   }
 
@@ -267,13 +309,13 @@ class DartBuilder {
         ..modifier = FieldModifier.final$
         ..type = refer(type)));
 
-      if (name.startsWith("_")) {
-        final argName = name.replaceFirst('_', "");
+      if (name.startsWith('_')) {
+        final argName = name.replaceFirst('_', '');
         constructor.optionalParameters.add(Parameter((p) => p
           ..name = argName
           ..type = refer(type)
           ..named = true));
-        constructor.initializers.add(Code("this.$name = $argName"));
+        constructor.initializers.add(Code('this.$name = $argName'));
       } else {
         constructor.optionalParameters.add(Parameter((p) => p
           ..name = name
@@ -285,13 +327,13 @@ class DartBuilder {
     result.methods.add(
       Method((b) => b
         ..name = 'getByKey'
-        ..returns = refer("String")
+        ..returns = refer('String')
         ..body = Code(_getLabelsCase(section.labels))
         ..requiredParameters.addAll([
           Parameter(
             (b) => b
-              ..type = refer("String")
-              ..name = "key",
+              ..type = refer('String')
+              ..name = 'key',
           )
         ])),
     );
@@ -300,58 +342,58 @@ class DartBuilder {
       if (label.templatedValues.isEmpty &&
           label.cases.length == 1 &&
           label.cases.first.condition is DefaultCondition) {
-        addField("String", label.normalizedKey);
+        addField('String', label.normalizedKey);
       } else {
         // If we have templated values, a function is generated
         final functionTypeName =
-            section.normalizedName + "_" + label.normalizedKey;
-        var functionArguments = "";
+            section.normalizedName + '_' + label.normalizedKey;
+        var functionArguments = '';
 
         final categories =
             label.cases.where((x) => x.condition is CategoryCondition);
         Category category;
         if (categories.isNotEmpty) {
           category = (categories.first.condition as CategoryCondition).category;
-          functionArguments += "${category.normalizedKey} condition";
+          functionArguments += '${category.normalizedKey} condition';
         }
 
         if (label.templatedValues.isNotEmpty) {
-          if (functionArguments.isNotEmpty) functionArguments += ",";
-          functionArguments += "{" +
+          if (functionArguments.isNotEmpty) functionArguments += ',';
+          functionArguments += '{' +
               label.templatedValues
-                  .map((x) => "@required ${x.type} ${x.normalizedKey}")
-                  .join(", ") +
-              "}";
+                  .map((x) => '@required ${x.type} ${x.normalizedKey}')
+                  .join(', ') +
+              '}';
         }
 
         _library.body.add(
-          Code("typedef String $functionTypeName($functionArguments);"),
+          Code('typedef String $functionTypeName($functionArguments);'),
         );
-        addField(functionTypeName, "_" + label.normalizedKey);
+        addField(functionTypeName, '_' + label.normalizedKey);
 
         final templatedCallArgs = label.templatedValues.isNotEmpty
             ? label.templatedValues
-                .map((x) => "${x.normalizedKey}: ${x.normalizedKey},")
+                .map((x) => '${x.normalizedKey}: ${x.normalizedKey},')
                 .join()
-            : "";
-        final conditionCallArgs = categories.isNotEmpty ? "condition, " : "";
+            : '';
+        final conditionCallArgs = categories.isNotEmpty ? 'condition, ' : '';
 
         // A method proxy is needed for @required arguments
         result.methods.add(
           Method(
             (b) => b
               ..name = label.normalizedKey
-              ..returns = refer("String")
+              ..returns = refer('String')
               ..lambda = true
               ..body = Code(
-                "this._${label.normalizedKey}($conditionCallArgs$templatedCallArgs)",
+                'this._${label.normalizedKey}($conditionCallArgs$templatedCallArgs)',
               )
               ..requiredParameters.addAll(categories.isNotEmpty
                   ? [
                       Parameter(
                         (b) => b
                           ..type = refer(category.normalizedKey)
-                          ..name = "condition",
+                          ..name = 'condition',
                       )
                     ]
                   : [])