+++ flutter_sheet_localization/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0
+
+- Added version to force regeneration.
+
 ## 1.0.0
 
 - Initial version